rfkill: Add hidden type and use it for OBDA8723 device
The HKC laptop is shipping a RTL8723BS WiFi/BT transceiver (APCI device
OBDA8723).

This transceiver needs to be enabled before starting to upload the
firmware. To enable the chip at boot we are using a downstream patch
taken from [0] where a new rfkill switch is created and driven at boot
to enable the BT before starting the systemd service used to upload the
firmware.

The problem with this approach is that this switch cannot really be used
to switch off the BT RF part (even though it supposed to do so), since
when doing so we are later unable to have it working fine again.

This is especially problematic because the BT GUI interface is actually
driving all the rfkill switches marked with type 'bluetooth' when
switching the BT off.

To avoid this problem we introduce a new 'hidden' type of rfkill switch
that we can use that to mark that specific switch. The net result is
that when BT is toggled using the GUI, only the rfkill switch related to
the UART interface is actually toggled.

[0] https://patchwork.kernel.org/patch/9671285/

Signed-off-by: Carlo Caione <carlo@endlessm.com>